Added capability to define the maximum number of particles that are integrated simultaneously in the configuration file.
Added capability to include a radial force (e.g., solar radiation
pressure) in the propagation of orbits (n-body only).

Minor correction to addMultinormalDeviate in SphericalCoordinates.

Numerous minor corrections to, e.g., covarianceSampling and
getNominalOrbit in StochasticOrbit.

Changed name for function testing the availability of a discrete PDF
estimate in StochasticOrbit from 'containsSampledPDF' to
'containsDiscretePDF'.

Changed name for function returning the discrete PDF estimate in
StochasticOrbit from 'getPDFValues' to 'getDiscretePDF'.

A major change in StochasticOrbit is that stochasticorbit%pdf_arr_cmp
will from now on always contain the correct pdf, i.e., based on
exp(chi2) for original ranging, based on repetitions for MCMC ranging,
and so forth.

Added capability to return covariance matrix for ephemerides also in
the case that only a discrete orbital-element PDF is available (and
not a covariance matrix).

Deleted virtualObservationMCMC routine and changed name for
virtualObservationMCMC2 to virtualObservationMCMC.

Changed the re-initialization of Keplerian orbits in simplexOrbit.

Added Position Angle (for the direction of motion) to the output
variables for tasks 'ephemeris' and 'obsplanner' in oorb.f90.