bugfix: identify JPL ephemeris type by scanning the filename, not path
JPL_ephemeris_init used to identify the ephemeris type by searching for
string XXX (where XXX=405, 406, 430, ..., in that order) in the full path
name using INDEX().  This could lead to misidentification of the ephemeris
type if the path name contained the substring XXX, that was different from
the actual ephemeris type. For example, placing de430.dat into
a directory named `openorb_15405` and running:

	OORB_DATA=openorb_15405/ python test.py

would cause an error as de430.dat would be read as if it was DE405.