Infer version information from git
Version information is now inferred from git tags, and compiled into both
the `oorb` binary (printable using `oorb --version`), and the `pyoorb`
module (in `pyoorb.__version__`). Generated versions are PEP440 compliant.

If the code is not being built in a git repository, whatever is stored in a
file named `VERSION` will be used. This allows for tarball distributions.

See the documentation in `build-tools/compute-version.sh` for the complete
version construction logic.